[radix cache split] Fold the insert-walk tests into the unified suite module
Moves the resumable-insert, returned-values-drain, SWA-lock, and
prefetch-ordering test classes into test_unified_radix_cache_unittest.py
and retires test_unified_insert_walk.py; the four cases that duplicated
suite originals (the write-back fallback pair and the FIFO contract pair)
are dropped in favor of those originals. The suite registration absorbs
the retired file's est_time.
